Moved DeleteVenue component to seperated jsx file
