+++ src/components/venue/ManageVenue.jsx
@@ -1,9 +1,10 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { TiArrowSortedDown } from 'react-icons/ti';
 import { AiOutlineEdit } from 'react-icons/ai';
-import { AiOutlineDelete } from 'react-icons/ai';
+
 import { AiOutlineClose } from 'react-icons/ai';
-import { getVenue } from '../../services/authorization/apiBase';
+
+import { DeleteVenue } from './DeleteVenue';
 
 /* import houseImg from '../../assets/img/house.jpg'; */
 
@@ -106,56 +107,6 @@ function VenueCreation() {
       </div>
       {isOpen && <VenueForm />}
     </div>
-  );
-}
-
-function DeleteVenue({ venueId, onError, onMessage }) {
-  const [successMessage, setSuccessMessage] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const handleDelete = async () => {
-    const accessToken = localStorage.getItem('accessToken');
-    try {
-      // Perform the delete request
-      // You can use the fetch API or any other library you prefer
-      const response = await fetch(getVenue + `${venueId}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-type': 'application/json',
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
-
-      if (response.ok) {
-        // Handle the successful deletion
-        // You can update the UI or perform any other necessary actions
-        console.log('Venue deleted successfully');
-        setSuccessMessage('Venue deleted successfully, page will refresh!');
-        setTimeout(() => {
-          setSuccessMessage('');
-        }, 3000);
-      } else {
-        // Handle the error case if the deletion was not successful
-        console.log('Failed to delete the venue');
-        setErrorMessage('Error deleting the venue');
-      }
-    } catch (error) {
-      // Handle any network or other errors
-      console.error('Error deleting the venue:', error);
-      setErrorMessage('Error deleting the venue', error);
-    }
-  };
-
-  useEffect(() => {
-    onError(errorMessage);
-    onMessage(successMessage);
-  }, [errorMessage, successMessage, onError, onMessage]);
-
-  return (
-    <>
-      <button type="button" onClick={handleDelete}>
-        <AiOutlineDelete className="icons-style_edit" />
-      </button>
-    </>
   );
 }
